backend/app/llm/llama_service.py
# filepath: d:\Vakeel\backend\app\llm\llama_service.py
import os
import logging
from typing import Dict, List, Optional, Any
from dotenv import load_dotenv
import random

load_dotenv()

logger = logging.getLogger(__name__)

class LlamaLegalService:
    def __init__(self):
        self.model = None
        self.tokenizer = None
        self.pipeline = None
        self.model_path = os.getenv("LLM_MODEL_PATH", "meta-llama/Llama-2-7b-chat-hf")
        
        # System prompt specifically for legal assistance
        self.system_prompt = """You are a legal assistant named Vakeel, designed to provide accurate
legal information and guidance. Please answer legal queries clearly, citing relevant laws and
regulations. Always clarify that you are providing information, not legal advice, and recommend
consulting with a qualified attorney for specific legal advice. Focus on providing factual,
educational information about legal concepts, rights, and procedures."""

        # Predefined responses for mock service
        self.mock_responses = [
            "Based on general legal principles, this situation would typically fall under contract law. "
            "However, I should note that I am providing information, not legal advice. For your specific "
            "situation, it would be best to consult with a qualified attorney.",
            
            "This appears to be a matter of property rights. In most jurisdictions, property laws provide "
            "protections for homeowners and tenants. I recommend consulting with a real estate attorney "
            "who can provide advice tailored to your specific circumstances.",
            
            "From a legal perspective, employment matters are governed by both federal and state laws. "
            "The specifics can vary significantly by jurisdiction. Please consider speaking with an employment "
            "law specialist for advice specific to your situation.",
            
            "Family law matters are often complex and emotionally challenging. While I can provide general "
            "information about legal concepts, a family law attorney would be best equipped to advise you "
            "on your specific rights and obligations.",
            
            "Criminal law questions require careful consideration of many factors. While I can explain general "
            "legal concepts, I must emphasize that I am providing information, not legal advice. Please consult "
            "with a criminal defense attorney for guidance specific to your situation."
        ]
        
        logger.info("Mock LLM service initialized")

    def load_model(self):
        """Mock loading the Llama2 model and tokenizer"""
        try:
            logger.info("Mock model loaded successfully")
            return True
        except Exception as e:
            logger.error(f"Error in mock model setup: {str(e)}")
            return False

    def generate_response(self, query: str) -> Dict:
        """Generate a mock legal response based on the user's query"""
        try:
            # Select a random predefined response
            response = random.choice(self.mock_responses)
            
            # Add some personalization based on the query
            category = self.categorize_query(query)
            if category != "general":
                response = f"Regarding your question about {category} law: {response}"
            
            return {
                "response": response,
                "error": None
            }
            
        except Exception as e:
            logger.error(f"Error generating mock response: {str(e)}")
            return {
                "response": "I apologize, but I encountered an error while processing your legal question. Please try again with a more specific query.",
                "error": str(e)
            }

    def categorize_query(self, query: str) -> str:
        """Attempt to categorize the legal query"""
        categories = {
            "criminal": ["arrest", "crime", "criminal", "offense", "police", "prosecution"],
            "civil": ["contract", "damages", "lawsuit", "sue", "civil case"],
            "family": ["divorce", "custody", "alimony", "child support", "marriage"],
            "property": ["real estate", "landlord", "tenant", "property", "eviction"],
            "employment": ["workplace", "fired", "termination", "salary", "employer"],
            "constitutional": ["rights", "freedom", "constitutional", "amendment"],
            "immigration": ["visa", "citizenship", "immigrant", "deportation", "asylum"],
        }
        
        query_lower = query.lower()
        
        for category, keywords in categories.items():
            for keyword in keywords:
                if keyword in query_lower:
                    return category
                    
        return "general"
