backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
sqlalchemy>=1.4.0,<1.5.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
bcrypt>=3.2.0
python-dotenv>=0.19.0
requests>=2.26.0
email-validator>=1.1.3